+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - fixed UI flickering
 - added "block_pose" property to set the pose used for blocking
 - backported glint-material settings
+- added Epic Fight Armor Compat